% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leman2000.R
\name{leman2000}
\alias{leman2000}
\title{Leman's (2000) Tonal Contextuality Model}
\usage{
leman2000(
  input_file,
  local_decay_sec,
  global_decay_sec,
  windows = NULL,
  windowing_function = mean,
  keep_auditory_nerve = FALSE,
  keep_periodicity_pitch = FALSE
)
}
\arguments{
\item{input_file}{(Character scalar)
Path to the input file (should be wav format, with a .wav extension).}

\item{local_decay_sec}{(Numeric)
Value of the local decay parameter.
If vectorised, then the function will output data for all combinations
of this parameter with the global_decay_sec parameter.}

\item{global_decay_sec}{(Numeric)
Value of the global decay parameter.
If vectorised, then the function will output data for all combinations
of this parameter with the local_decay_parameter parameter.}

\item{windows}{(List, optional)
Optional specification of time windows to perform averaging over.
Each element of the list corresponds to a time window,
and should be a numeric vector of length 2, providing the window start and
end time in seconds. Averaging is performed over all timepoints greater
or equal to the window start and less than the window end.}

\item{windowing_function}{(Function)
This function is used to average within windows. By default the mean
is used but other functions (e.g. the median) could be provided instead.}

\item{keep_auditory_nerve}{(Boolean)
The model can output auditory nerve simulation outputs, but these
are omitted by default because they take up a lot of space.}

\item{keep_periodicity_pitch}{(Boolean)
The model can output auditory periodicity pitch outputs, but these
are omitted by default because they take up a lot of space.}
}
\value{
Returns a list with several elements:

\code{audio_length_sec} gives the length of the audio file in seconds.

\code{num_channels} gives the number of channels of the input audio file.

\code{sample_rate} gives the sample rate of the audio file.

\code{local_global_comparison} is a tibble giving running correlations
between local and global images over the time course of the audio files.

\code{windowed_local_global_comparison} is a tibble giving windowed averages
of these local-global images for the specified time windows.

\code{keep_auditory_nerve} provides the auditory nerve images if requested.

\code{periodicity_pitch} provides the periodicity_pitch images if requested.
}
\description{
This model was published in a 2000 Music Perception paper, and was shown
to provide a psychoacoustic account of the Krumhansl-Kessler probe-tone data.
}
